refactor(e2e): convert token + blacklist tests to automine harnesses
Make TokenContractTest and BlacklistTokenContractTest extend AutomineTestContext:
they keep their TokenSimulator, opt-in base/mint snapshots, and the Role helper, and
run their domain setup after super.setup(). crossTimestampOfChange now delegates to
the base markProvenAndWarp, and the harness loggers move to the e2e:automine:token /
e2e:automine:blacklist namespaces. Move the 17 token/blacklist test files plus the two
harnesses under automine/token/ (blacklist files gain a blacklist_ prefix to avoid
collisions), switching each from setup({ ...AUTOMINE_E2E_OPTS }) to the base default
and naming each describe after its path. Files with more than one top-level it gain
the .parallel suffix.

Adds the automine/token glob to both bootstrap.sh arrays, repoints the access_control
flake entry, and updates the e2e_persistence harness import.
