+++ yarn-project/end-to-end/src/automine/token/blacklist_shielding.parallel.test.ts
@@ -1,19 +1,18 @@
 import { computeSecretHash } from '@aztec/aztec.js/crypto';
 import { Fr } from '@aztec/aztec.js/fields';
 
-import { AUTOMINE_E2E_OPTS } from '../fixtures/fixtures.js';
-import { U128_UNDERFLOW_ERROR } from '../fixtures/index.js';
+import { U128_UNDERFLOW_ERROR } from '../../fixtures/index.js';
 import { BlacklistTokenContractTest } from './blacklist_token_contract_test.js';
 
 // Covers the shield (public→private) and redeem_shield operations on TokenBlacklist, including
 // authwit-delegated shielding and blacklist enforcement. Setup: single node with AutomineSequencer,
 // 3 accounts, initial mint applied. Time-warp required during setup to cross role-change delay.
-describe('e2e_blacklist_token_contract shield + redeem_shield', () => {
+describe('automine/token/blacklist_shielding', () => {
   const t = new BlacklistTokenContractTest('shield');
   let { asset, tokenSim, wallet, adminAddress, otherAddress, blacklistedAddress } = t;
 
   beforeAll(async () => {
-    await t.setup({ ...AUTOMINE_E2E_OPTS });
+    await t.setup();
     await t.applyMint(); // Beware that we are adding the admin as minter here
     // Have to destructure again to ensure we have latest refs.
     ({ asset, tokenSim, wallet, adminAddress, otherAddress, blacklistedAddress } = t);
