+++ yarn-project/end-to-end/src/automine/token/blacklist_unshielding.parallel.test.ts
@@ -1,19 +1,19 @@
 import { computeAuthWitMessageHash } from '@aztec/aztec.js/authorization';
 import { Fr } from '@aztec/aztec.js/fields';
 
-import { sendThroughAuthwitProxy, simulateThroughAuthwitProxy } from '../fixtures/authwit_proxy.js';
-import { AUTOMINE_E2E_OPTS, DUPLICATE_NULLIFIER_ERROR } from '../fixtures/fixtures.js';
+import { sendThroughAuthwitProxy, simulateThroughAuthwitProxy } from '../../fixtures/authwit_proxy.js';
+import { DUPLICATE_NULLIFIER_ERROR } from '../../fixtures/fixtures.js';
 import { BlacklistTokenContractTest } from './blacklist_token_contract_test.js';
 
 // Covers the unshield (private→public) operation on TokenBlacklist, including authwit-delegated unshielding
 // and blacklist enforcement on sender and recipient. Setup: single node with AutomineSequencer, 3 accounts,
 // initial mint applied. Time-warp required during setup to cross role-change delay.
-describe('e2e_blacklist_token_contract unshielding', () => {
+describe('automine/token/blacklist_unshielding', () => {
   const t = new BlacklistTokenContractTest('unshielding');
   let { asset, tokenSim, wallet, adminAddress, otherAddress, blacklistedAddress } = t;
 
   beforeAll(async () => {
-    await t.setup({ ...AUTOMINE_E2E_OPTS });
+    await t.setup();
     // Beware that we are adding the admin as minter here, which is very slow because it needs multiple blocks.
     await t.applyMint();
     // Have to destructure again to ensure we have latest refs.
